Moving some of the modules around to have more consistant naming and remove deplicate code
