+++ graphbuild/builder.go
@@ -1,4 +1,4 @@
-package graph
+package graphbuild
 
 import (
 	"fmt"
@@ -7,11 +7,12 @@ import (
 	"path/filepath"
 
 	"github.com/bmeg/grip/gripql"
+	"github.com/bmeg/sifter/loader"
 	"github.com/bmeg/sifter/schema"
 )
 
 type DomainClassInfo struct {
-	emitter   Emitter
+	emitter   loader.GraphEmitter
 	gc        *Check
 	om        *ObjectMap
 	vertCount int64
@@ -19,14 +20,14 @@ type DomainClassInfo struct {
 }
 
 type DomainInfo struct {
-	emitter Emitter
+	emitter loader.GraphEmitter
 	gc      *Check
 	dm      *DomainMap
 	classes map[string]*DomainClassInfo
 }
 
 type Builder struct {
-	emitter Emitter
+	emitter loader.GraphEmitter
 	sc      schema.Schemas
 	gm      *Mapping
 	gc      *Check
@@ -34,7 +35,7 @@ type Builder struct {
 }
 
 func NewBuilder(driver string, sc schema.Schemas, workdir string) (*Builder, error) {
-	emitter, err := NewGraphEmitter(driver)
+	emitter, err := loader.NewGraphEmitter(driver)
 	if err != nil {
 		return nil, err
 	}
@@ -99,7 +100,7 @@ func (b *Builder) Process(prefix string, class string, in chan map[string]interf
 	}
 }
 
-func (b *Builder) GenerateGraph(objMap *ObjectMap, class string, data map[string]interface{}, emitter Emitter) error {
+func (b *Builder) GenerateGraph(objMap *ObjectMap, class string, data map[string]interface{}, emitter loader.GraphEmitter) error {
 	if o, err := b.sc.Generate(class, data); err == nil {
 		for _, j := range o {
 			if j.Vertex != nil {
