+++ graphbuild/graphCheck.go
@@ -1,4 +1,4 @@
-package graph
+package graphbuild
 
 import (
 	"log"
