+++ loader/dir.go
@@ -1,4 +1,4 @@
-package emitter
+package loader
 
 import (
 	"compress/gzip"
@@ -13,24 +13,89 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/bmeg/grip/gripql"
 	"github.com/bmeg/sifter/schema"
+	"github.com/golang/protobuf/jsonpb"
 )
 
 type DirEmitter struct {
+	jm      jsonpb.Marshaler
 	dir     string
 	mux     sync.Mutex
 	schemas *schema.Schemas
-	dout    map[string]io.WriteCloser
+	vout    map[string]io.WriteCloser
+	eout    map[string]io.WriteCloser
 	oout    map[string]io.WriteCloser
+	dout    map[string]io.WriteCloser
 }
 
 func NewDirEmitter(dir string, schemas *schema.Schemas) *DirEmitter {
 	log.Printf("Emitting to %s", dir)
 	return &DirEmitter{
+		jm:      jsonpb.Marshaler{},
 		dir:     dir,
-		schemas: schemas,
-		dout:    map[string]io.WriteCloser{},
+		vout:    map[string]io.WriteCloser{},
+		eout:    map[string]io.WriteCloser{},
 		oout:    map[string]io.WriteCloser{},
+		dout:    map[string]io.WriteCloser{},
+		schemas: schemas,
+	}
+}
+
+func (s *DirEmitter) EmitVertex(v *gripql.Vertex) error {
+	s.mux.Lock()
+	defer s.mux.Unlock()
+	f, ok := s.vout[v.Label]
+	if !ok {
+		j, err := os.Create(path.Join(s.dir, v.Label+".Vertex.json.gz"))
+		if err != nil {
+			log.Printf("Error: %s", err)
+			return err
+		}
+		f = gzip.NewWriter(j)
+		s.vout[v.Label] = f
+	}
+	o, _ := s.jm.MarshalToString(v)
+	f.Write([]byte(o))
+	f.Write([]byte("\n"))
+	return nil
+}
+
+func (s *DirEmitter) EmitEdge(e *gripql.Edge) error {
+	s.mux.Lock()
+	defer s.mux.Unlock()
+	f, ok := s.eout[e.Label]
+	if !ok {
+		j, err := os.Create(path.Join(s.dir, e.Label+".Edge.json.gz"))
+		if err != nil {
+			return err
+		}
+		f = gzip.NewWriter(j)
+		s.eout[e.Label] = f
+	}
+	o, err := s.jm.MarshalToString(e)
+	if err != nil {
+		log.Printf("Error: %s", err)
+		return err
+	}
+	f.Write([]byte(o))
+	f.Write([]byte("\n"))
+	return nil
+}
+
+func (s *DirEmitter) Close() {
+	log.Printf("Closing emitter")
+	for _, v := range s.vout {
+		v.Close()
+	}
+	for _, v := range s.eout {
+		v.Close()
+	}
+	for _, v := range s.oout {
+		v.Close()
+	}
+	for _, v := range s.dout {
+		v.Close()
 	}
 }
 
@@ -74,16 +139,6 @@ func (s *DirEmitter) EmitObject(prefix string, objClass string, i map[string]int
 	f.Write([]byte(o))
 	f.Write([]byte("\n"))
 	return nil
-}
-
-func (s *DirEmitter) Close() {
-	log.Printf("Closing dir emitter")
-	for _, v := range s.oout {
-		v.Close()
-	}
-	for _, v := range s.dout {
-		v.Close()
-	}
 }
 
 type dirTableEmitter struct {
